refactor: apply PR feedbacks
