+++ src/helpers/calculator.ts
@@ -108,106 +108,3 @@ export function getTextWidth(text: string, font: string) {
 
   return ctx.measureText(text).width;
 }
-
-function limitNegative(value: number, min: number, max: number) {
-  if (value >= max) {
-    return 0;
-  }
-
-  let result = value;
-
-  if (max < 0) {
-    result = Math.min(value - max, 0);
-  }
-
-  if (value < min) {
-    result -= value - min;
-  }
-
-  return result;
-}
-
-function limitPositive(value: number, min: number, max: number) {
-  let result = value;
-
-  if (min > 0) {
-    result = Math.max(value - min, 0);
-  }
-
-  if (value > max) {
-    result -= value - max;
-  }
-
-  return result;
-}
-
-export function calibrateBoxDrawingValue(value: number, min: number, max: number) {
-  return value < 0 ? limitNegative(value, min, max) : limitPositive(value, min, max);
-}
-
-export function sumOfPrevValues(values: number[], currentIndex: number, included = false) {
-  const curValue = values[currentIndex];
-
-  return values.reduce((total, value, idx) => {
-    const isPrev = included ? idx <= currentIndex : idx < currentIndex;
-    const isSameSign = value * curValue >= 0;
-
-    if (isPrev && isSameSign) {
-      return total + value;
-    }
-
-    return total;
-  }, 0);
-}
-
-export function exceedEdge(values: number[], currentIndex: number, min: number, max: number) {
-  const value = values[currentIndex];
-  const totalOfPrevValues = sumOfPrevValues(values, currentIndex, false);
-  const positive = value >= 0;
-  const negative = value < 0;
-
-  return (
-    (positive && totalOfPrevValues >= max) ||
-    (negative && totalOfPrevValues <= min) ||
-    (currentIndex === 0 && positive && value < min) ||
-    (currentIndex === 0 && negative && value > max)
-  );
-}
-
-export function calibrateBoxStackDrawingValue(
-  values: number[],
-  currentIndex: number,
-  min: number,
-  max: number
-) {
-  const totalOfPrevValues = sumOfPrevValues(values, currentIndex, false);
-  const totalOfValues = sumOfPrevValues(values, currentIndex, true);
-
-  if (exceedEdge(values, currentIndex, min, max)) {
-    return;
-  }
-
-  if (currentIndex === 0) {
-    return calibrateBoxDrawingValue(values[currentIndex], min, max);
-  }
-
-  let result = values[currentIndex];
-
-  if (totalOfValues > max) {
-    result = max - totalOfPrevValues;
-  }
-
-  if (totalOfValues < min) {
-    result = min - totalOfPrevValues;
-  }
-
-  if (totalOfPrevValues < min) {
-    result = totalOfValues - min;
-  }
-
-  if (totalOfPrevValues > max) {
-    result = totalOfValues - max;
-  }
-
-  return result;
-}
